+++ contracts/libs/crypto/Schnorr256.sol
@@ -7,7 +7,8 @@ import {MemoryUtils} from "../utils/MemoryUtils.sol";
 /**
  * @notice Cryptography module
  *
- * This library provides functionality for Schnorr signature verification over any 256-bit curve.
+ * This library provides functionality for Schnorr signature verification over any 256-bit curve,
+ * and secret extraction from a standard/adaptor Schnorr signature pair.
  */
 library Schnorr256 {
     using MemoryUtils for *;
@@ -15,6 +16,8 @@ library Schnorr256 {
 
     error LengthIsNot64();
     error LengthIsNot96();
+    error InvalidSignature();
+    error InvalidAdaptorSignature();
 
     /**
      * @notice The function to verify the Schnorr signature
@@ -22,6 +25,7 @@ library Schnorr256 {
      * @param hashedMessage_ the already hashed message to be verified.
      * @param signature_ the Schnorr signature. Equals to `bytes(R) + bytes(e)`.
      * @param pubKey_ the full public key of a signer. Equals to `bytes(x) + bytes(y)`.
+     * @return True if the signature is valid, false otherwise.
      */
     function verify(
         EC256.Curve memory ec,
@@ -29,6 +33,70 @@ library Schnorr256 {
         bytes memory signature_,
         bytes memory pubKey_
     ) internal view returns (bool) {
+        (EC256.APoint memory r_, ) = _parseSignature(signature_);
+
+        return _verify(ec, hashedMessage_, signature_, pubKey_, r_);
+    }
+
+    /**
+     * @notice The function to extract the adaptor secret from a pair of Schnorr signatures.
+     * @dev Reverts if the standard or adaptor Schnorr signature is invalid.
+     *
+     *      The adaptor Schnorr signature is expected to be computed as:
+     *          c = H(P || R + T || m)
+     *          e' = (r + c * privKey) mod n
+     *          signature = (R, e')
+     *
+     *      The standard Schnorr signature is expected to be computed from the adaptor one as:
+     *          e = e' + t = (r + t + c * privKey) mod n
+     *          signature = (R + T, e)
+     *
+     *      Secret extraction is performed as follows:
+     *          t = (e - e') mod n
+     *
+     * @param ec the 256-bit curve parameters.
+     * @param hashedMessage_ the already hashed message signed.
+     * @param signature_ the Schnorr signature. Equals to `bytes(R + T) + bytes(e)`.
+     * @param adaptorSignature_ the adaptor Schnorr signature. Equals to `bytes(R) + bytes(e')`.
+     * @param pubKey_ the full public key of a signer. Equals to `bytes(x) + bytes(y)`.
+     * @return secret_ the secret scalar that was used in the adaptor signature.
+     */
+    function extract(
+        EC256.Curve memory ec,
+        bytes32 hashedMessage_,
+        bytes memory signature_,
+        bytes memory adaptorSignature_,
+        bytes memory pubKey_
+    ) internal view returns (uint256 secret_) {
+        (EC256.APoint memory nonce_, uint256 sigScalar_) = _parseSignature(signature_);
+        (EC256.APoint memory adaptorNonce_, uint256 adaptorScalar_) = _parseSignature(
+            adaptorSignature_
+        );
+
+        secret_ = addmod(sigScalar_, ec.n - adaptorScalar_, ec.n);
+
+        if (!_verify(ec, hashedMessage_, signature_, pubKey_, nonce_)) {
+            revert InvalidSignature();
+        }
+
+        EC256.APoint memory secretPoint_ = ec.toAffine(ec.jMultShamir(ec.jbasepoint(), secret_));
+
+        EC256.APoint memory rt_ = ec.toAffine(
+            ec.jAddPoint(adaptorNonce_.toJacobian(), secretPoint_.toJacobian())
+        );
+
+        if (!_verify(ec, hashedMessage_, adaptorSignature_, pubKey_, rt_)) {
+            revert InvalidAdaptorSignature();
+        }
+    }
+
+    function _verify(
+        EC256.Curve memory ec,
+        bytes32 hashedMessage_,
+        bytes memory signature_,
+        bytes memory pubKey_,
+        EC256.APoint memory RT_
+    ) private view returns (bool) {
         (EC256.APoint memory r_, uint256 e_) = _parseSignature(signature_);
         EC256.APoint memory p_ = _parsePubKey(pubKey_);
 
@@ -39,7 +107,7 @@ library Schnorr256 {
         EC256.JPoint memory lhs_ = ec.jMultShamir(ec.jbasepoint(), e_);
 
         uint256 c_ = ec.toScalar(
-            uint256(keccak256(abi.encodePacked(ec.gx, ec.gy, r_.x, r_.y, hashedMessage_)))
+            uint256(keccak256(abi.encodePacked(p_.x, p_.y, RT_.x, RT_.y, hashedMessage_)))
         );
 
         EC256.JPoint memory rhs_ = ec.jMultShamir(p_.toJacobian(), c_);