+++ test/libs/crypto/Schnorr256.test.ts
@@ -25,7 +25,7 @@ describe("Schnorr256", () => {
     return ethers.solidityPacked(["uint256", "uint256"], [p.x, p.y]);
   };
 
-  const schnorrSign = function (hashedMessage: string, privKey: bigint) {
+  const schnorrSign = function (hashedMessage: string, privKey: bigint, pubKey: AffinePoint<bigint>) {
     const randomness = schnorrKeyPair();
     const k = randomness.privKey;
     const R = randomness.pubKey;
@@ -33,7 +33,7 @@ describe("Schnorr256", () => {
     const c = BigInt(
       ethers.solidityPackedKeccak256(
         ["uint256", "uint256", "uint256", "uint256", "bytes32"],
-        [secp256k1.CURVE.Gx, secp256k1.CURVE.Gy, R.x, R.y, hashedMessage],
+        [pubKey.x, pubKey.y, R.x, R.y, hashedMessage],
       ),
     );
     const e = (k + c * privKey) % secp256k1.CURVE.n;
@@ -41,12 +41,38 @@ describe("Schnorr256", () => {
     return ethers.solidityPacked(["uint256", "uint256", "uint256"], [R.x, R.y, e]);
   };
 
+  const schnorrAdaptorSign = function (
+    hashedMessage: string,
+    privKey: bigint,
+    pubKey: AffinePoint<bigint>,
+    T: AffinePoint<bigint>,
+  ) {
+    const randomness = schnorrKeyPair();
+    const k = randomness.privKey;
+    const R = randomness.pubKey;
+
+    const RT = secp256k1.ProjectivePoint.fromAffine(R).add(secp256k1.ProjectivePoint.fromAffine(T)).toAffine();
+
+    const c = BigInt(
+      ethers.solidityPackedKeccak256(
+        ["uint256", "uint256", "uint256", "uint256", "bytes32"],
+        [pubKey.x, pubKey.y, RT.x, RT.y, hashedMessage],
+      ),
+    );
+
+    const e = (k + c * privKey) % secp256k1.CURVE.n;
+
+    const signature = ethers.solidityPacked(["uint256", "uint256", "uint256"], [R.x, R.y, e]);
+
+    return { signature, RT, e, T };
+  };
+
   const prepareParameters = function (message: string) {
     const { privKey, pubKey } = schnorrKeyPair();
 
     const hashedMessage = ethers.keccak256(message);
 
-    const signature = schnorrSign(hashedMessage, privKey);
+    const signature = schnorrSign(hashedMessage, privKey, pubKey);
 
     return {
       hashedMessage,
@@ -106,6 +132,63 @@ describe("Schnorr256", () => {
 
       const wrongPubKey = "0x" + ethers.toBeHex(0, 0x40).slice(2);
       expect(await schnorr.verifySECP256k1(hashedMessage, signature, wrongPubKey)).to.be.false;
+    });
+  });
+
+  describe("extract", () => {
+    it("should extract secret from two signatures correctly", async () => {
+      const { privKey, pubKey } = schnorrKeyPair();
+
+      const hashedMessage = ethers.keccak256("0x1337");
+
+      const t = bytesToNumberBE(secp256k1.utils.randomPrivateKey());
+      const T = secp256k1.ProjectivePoint.BASE.multiply(t).toAffine();
+
+      const { signature: adaptorSignature, RT, e } = schnorrAdaptorSign(hashedMessage, privKey, pubKey, T);
+
+      const signatureScalar = (e + t) % secp256k1.CURVE.n;
+
+      const signature = ethers.solidityPacked(["uint256", "uint256", "uint256"], [RT.x, RT.y, signatureScalar]);
+
+      expect(
+        await schnorr.extractSECP256k1(hashedMessage, signature, adaptorSignature, serializePoint(pubKey)),
+      ).to.be.eq(t);
+    });
+
+    it("should revert if signature is incorrect", async () => {
+      const { privKey, pubKey } = schnorrKeyPair();
+
+      const hashedMessage = ethers.keccak256("0x1227");
+
+      const t = bytesToNumberBE(secp256k1.utils.randomPrivateKey());
+      const T = secp256k1.ProjectivePoint.BASE.multiply(t).toAffine();
+
+      const { signature } = schnorrAdaptorSign(hashedMessage, privKey, pubKey, T);
+
+      await expect(
+        schnorr.extractSECP256k1(hashedMessage, signature, signature, serializePoint(pubKey)),
+      ).to.be.revertedWithCustomError(schnorr, "InvalidSignature");
+    });
+
+    it("should revert if adaptor signature is incorrect", async () => {
+      const { privKey, pubKey } = schnorrKeyPair();
+
+      const hashedMessage = ethers.keccak256("0x1337");
+
+      const t = bytesToNumberBE(secp256k1.utils.randomPrivateKey());
+      const T = secp256k1.ProjectivePoint.BASE.multiply(t).toAffine();
+      const T2 = secp256k1.ProjectivePoint.BASE.multiply(t - 2n).toAffine();
+
+      const { RT, e } = schnorrAdaptorSign(hashedMessage, privKey, pubKey, T);
+      const { signature: invalidAdaptorSignature } = schnorrAdaptorSign(hashedMessage, privKey, pubKey, T2);
+
+      const signatureScalar = (e + t) % secp256k1.CURVE.n;
+
+      const signature = ethers.solidityPacked(["uint256", "uint256", "uint256"], [RT.x, RT.y, signatureScalar]);
+
+      await expect(
+        schnorr.extractSECP256k1(hashedMessage, signature, invalidAdaptorSignature, serializePoint(pubKey)),
+      ).to.be.revertedWithCustomError(schnorr, "InvalidAdaptorSignature");
     });
   });
 });