+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [patch]
+
 - Added `extract` function to the `Schnorr256` library to extract a secret from a standard/adaptor Schnorr signature pair.
 
 ## [3.2.6]