+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [3.2.6]
 
+- Added `extract` function to the `Schnorr256` library to extract a secret from a standard/adaptor Schnorr signature pair.
+
+## [3.2.6]
+
 - Enhanced the `ABridge` contract architecture.
   - Adding modular handler contracts for assets and messages management.
   - Added `Batcher` contract for batched execution during redemption.